Described `kit configure`
As this exercise is all about capturing the behavior of the
existing script, I have done exactly that for this feature.

One might expect the "no tools" scenario to return an error,
but that is not what `c3tk` currently does, and we have
retained the upstream behavior in this description.